+++ src/main/scala/com/atomist/rug/runtime/TestDescriptor.scala
@@ -1,0 +1,8 @@
+package com.atomist.rug.runtime
+
+/**
+  * Used to describe tests on a Rug
+  * @param kind
+  * @param description
+  */
+case class TestDescriptor (kind: String, description: String)